Completed use-auto-controlled-state hook
