--- packages/react-components/src/popper-trigger/index.js
@@ -1,1 +1,0 @@
-export * from "./src";
