--- packages/react-components/src/popper-trigger/src/index.js
@@ -1,1 +1,0 @@
-export * from "./popper-trigger";
