--- packages/react-components/src/popper-trigger/src/popper-trigger.jsx
@@ -1,3 +1,0 @@
-export function PopperTrigger() {
-
-}
