+++ packages/react-components/src/shared/src/auto-controlled-state/index.js
@@ -1,1 +1,2 @@
 export * from "./auto-controlled-state";
+export * from "./use-auto-controlled-state";
